+++ refactor-me/DataHelpers/Helpers.cs
@@ -6,28 +6,10 @@ using System.IO;
 using System.Reflection;
 using System.Web;
 
-namespace Refactor_me.Models
+namespace Refactor_me.DataHelpers
 {
-    public class Helpers
+    public static class Helpers
     {
-        public static void AddParameter(IDbCommand command, string name, object value)
-        {
-            if (command == null)
-            {
-                throw new ArgumentNullException("command");
-            }
-
-            if (string.IsNullOrEmpty(name))
-            {
-                throw new ArgumentNullException("name");
-            }
-
-            var parameter = command.CreateParameter();
-            parameter.ParameterName = name;
-            parameter.Value = value ?? DBNull.Value;
-            command.Parameters.Add(parameter);
-        }
-
         public static IDbConnection NewConnection()
         {
             if (HttpContext.Current != null)
